requirements.txt
fastapi
uvicorn
aiosqlite
PyYAML
httpx
